requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cryptography==46.0.7
distro==1.9.0
fastapi==0.136.1
flatbuffers==25.12.19
grpcio==1.80.0
grpcio-tools==1.80.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.13
jiter==0.14.0
networkx==3.6.1
numpy==2.4.4
onnxruntime==1.25.0
openai==2.32.0
packaging==26.1
pdfminer.six==20231228
portalocker==2.10.1
protobuf==6.33.6
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
PyMuPDF==1.27.2.2
pymupdf-layout==1.27.2.2
pymupdf4llm==1.27.2.2
PyMuPDFb==1.24.3
python-dotenv==1.0.1
PyYAML==6.0.3
qdrant-client==1.9.1
sniffio==1.3.1
starlette==1.0.0
tabulate==0.10.0
tqdm==4.66.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
